+++ lithops/constants.py
@@ -19,8 +19,9 @@ import sys
 import tempfile
 
 LOGGER_LEVEL = 'info'
-LOGGER_FORMAT = ("%(asctime)s [%(levelname)s] %(name)s -- %(message)s")
-LOGGER_FORMAT_SHORT = ("[%(levelname)s] %(filename)s -- %(message)s")
+LOGGER_STREAM = 'ext://sys.stderr'
+LOGGER_FORMAT = "%(asctime)s [%(levelname)s] %(name)s -- %(message)s"
+LOGGER_FORMAT_SHORT = "[%(levelname)s] %(filename)s -- %(message)s"
 LOGGER_LEVEL_CHOICES = ["debug", "info", "warning", "error", "critical"]
 
 STORAGE_CLI_MSG = '{} Storage client created'