+++ lithops/serverless/backends/aliyun_fc/config.py
@@ -28,6 +28,7 @@ RUNTIME_TIMEOUT_MAX = 600        # Platform maximum
 RUNTIME_MEMORY_DEFAULT = 256
 RUNTIME_MEMORY_MAX = 3072
 MAX_CONCURRENT_WORKERS = 300
+INVOKE_POOL_THREADS_DEFAULT = 500
 
 CONNECTION_POOL_SIZE = 30
 
@@ -55,12 +56,23 @@ tblib
 
 def load_config(config_data=None):
 
+    if 'aliyun_fc' not in config_data:
+        raise Exception("aliyun_fc section is mandatory in the configuration")
+
+    required_parameters = ('public_endpoint', 'access_key_id', 'access_key_secret')
+
+    if set(required_parameters) > set(config_data['aliyun_fc']):
+        raise Exception('You must provide {} to access to Aliyun Function Compute '
+                        .format(required_parameters))
+
     this_version_str = version_str(sys.version_info)
     if this_version_str != '3.6':
         raise Exception('The functions backend Aliyun Function Compute currently'
                         ' only supports Python version 3.6.X and the local Python'
                         'version is {}'.format(this_version_str))
 
+    if 'runtime' in config_data['aliyun_fc']:
+        config_data['serverless']['runtime'] = config_data['aliyun_fc']['runtime']
     if 'runtime' not in config_data['serverless']:
         config_data['serverless']['runtime'] = 'default'
 
@@ -82,11 +94,6 @@ def load_config(config_data=None):
     else:
         config_data['lithops']['workers'] = MAX_CONCURRENT_WORKERS
 
-    if 'aliyun_fc' not in config_data:
-        raise Exception("aliyun_fc section is mandatory in the configuration")
-
-    required_parameters = ('public_endpoint', 'access_key_id', 'access_key_secret')
-
-    if set(required_parameters) > set(config_data['aliyun_fc']):
-        raise Exception('You must provide {} to access to Aliyun Function Compute '
-                        .format(required_parameters))
+    if 'invoke_pool_threads' not in config_data['aliyun_fc']:
+        config_data['aliyun_fc']['invoke_pool_threads'] = INVOKE_POOL_THREADS_DEFAULT
+    config_data['serverless']['invoke_pool_threads'] = config_data['aliyun_fc']['invoke_pool_threads']