+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## [v2.2.15.dev0]
+## [v2.2.17.dev0]
+
+### Added
+- [Config] Allow 'log_level' and 'log_format' keys in configuration
+- [Config] Allow 'log_stream' and 'log_filename' keys in configuration
+- [Config] Allow 'runtime' being configured at serverless backend level
+- [Config] Allow 'invoke_pool_threads' being configured at serverless backend level
+
+### Changed
+- [Core] Renamed utils.setup_logger() method to utils.setup_lithops_logger()
+
+
+## [v2.2.15]
 
 ### Added
 - [Joblib] Joblib backend upgraded